Removed remaining time from the UI, rounded the distance to 2 decimal places

+++ apps/Traffic/src/in/beetroute/apps/traffic/Route.java
@@ -31,7 +31,7 @@ public class Route {
 	// and so gson couldn't convert it to json.  (ummm, i think)
 	protected ArrayList<SimpleGeoPoint> points;
 	
-	public int drivingDistanceMeters = 0;
+	public float drivingDistanceMeters = 0;
 	public int estimatedTimeSeconds = 0;
 
 	public Route(ServiceProviders service) {
